+++ src/components/split-panel/split-panel.test.ts
@@ -185,7 +185,7 @@ describe('<wa-split-panel>', () => {
       describe('panel sizing vertical', () => {
         it('has two evenly sized panels by default', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel vertical style="height: 400px;">
+            html`<wa-split-panel orientation="vertical" style="height: 400px;">
               <div slot="start" data-testid="start-panel">Start</div>
               <div slot="end" data-testid="end-panel">End</div>
             </wa-split-panel>`,
@@ -199,7 +199,7 @@ describe('<wa-split-panel>', () => {
 
         it('changes the sizing of the panels based on the position attribute', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel position="25" vertical style="height: 400px;">
+            html`<wa-split-panel position="25" orientation="vertical" style="height: 400px;">
               <div slot="start" data-testid="start-panel">Start</div>
               <div slot="end" data-testid="end-panel">End</div>
             </wa-split-panel>`,
@@ -213,7 +213,7 @@ describe('<wa-split-panel>', () => {
 
         it('updates the position in pixels to the correct result', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel position="25" vertical style="height: 400px;">
+            html`<wa-split-panel position="25" orientation="vertical" style="height: 400px;">
               <div slot="start" data-testid="start-panel">Start</div>
               <div slot="end" data-testid="end-panel">End</div>
             </wa-split-panel>`,
@@ -228,7 +228,7 @@ describe('<wa-split-panel>', () => {
 
         it('emits the wa-reposition	event on position change ', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel vertical style="height: 400px;">
+            html`<wa-split-panel orientation="vertical" style="height: 400px;">
               <div slot="start">Start</div>
               <div slot="end">End</div>
             </wa-split-panel>`,
@@ -241,7 +241,7 @@ describe('<wa-split-panel>', () => {
 
         it('can be resized using the mouse ', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel vertical style="height: 400px;">
+            html`<wa-split-panel orientation="vertical" style="height: 400px;">
               <div slot="start">Start</div>
               <div slot="end">End</div>
             </wa-split-panel>`,
@@ -259,7 +259,7 @@ describe('<wa-split-panel>', () => {
 
         it('cannot be resized if disabled', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel disabled vertical style="height: 400px;">
+            html`<wa-split-panel disabled orientation="vertical" style="height: 400px;">
               <div slot="start">Start</div>
               <div slot="end">End</div>
             </wa-split-panel>`,
@@ -277,7 +277,7 @@ describe('<wa-split-panel>', () => {
 
         it('snaps to predefined positions', async () => {
           const splitPanel = await fixture<WaSplitPanel>(
-            html`<wa-split-panel vertical style="height: 400px;">
+            html`<wa-split-panel orientation="vertical" style="height: 400px;">
               <div slot="start">Start</div>
               <div slot="end">End</div>
             </wa-split-panel>`,