+++ src/components/split-panel/split-panel.ts
@@ -55,8 +55,8 @@ export default class WaSplitPanel extends WebAwesomeElement {
   /** The current position of the divider from the primary panel's edge in pixels. */
   @property({ attribute: 'position-in-pixels', type: Number }) positionInPixels: number;
 
-  /** Draws the split panel in a vertical orientation with the start and end panels stacked. */
-  @property({ type: Boolean, reflect: true }) vertical = false;
+  /** Sets the split panel's orientation. */
+  @property({ reflect: true }) orientation: 'horizontal' | 'vertical' = 'horizontal';
 
   /** Disables resizing. Note that the position may still change as a result of resizing the host element. */
   @property({ type: Boolean, reflect: true }) disabled = false;
@@ -93,7 +93,7 @@ export default class WaSplitPanel extends WebAwesomeElement {
 
   private detectSize() {
     const { width, height } = this.getBoundingClientRect();
-    this.size = this.vertical ? height : width;
+    this.size = this.orientation === 'vertical' ? height : width;
   }
 
   private percentageToPixels(value: number) {
@@ -118,7 +118,7 @@ export default class WaSplitPanel extends WebAwesomeElement {
 
     drag(this, {
       onMove: (x, y) => {
-        let newPositionInPixels = this.vertical ? y : x;
+        let newPositionInPixels = this.orientation === 'vertical' ? y : x;
 
         // Flip for end panels
         if (this.primary === 'end') {
@@ -138,7 +138,7 @@ export default class WaSplitPanel extends WebAwesomeElement {
               snapPoint = parseFloat(value);
             }
 
-            if (isRtl && !this.vertical) {
+            if (isRtl && this.orientation === 'horizontal') {
               snapPoint = this.size - snapPoint;
             }
 
@@ -168,11 +168,17 @@ export default class WaSplitPanel extends WebAwesomeElement {
 
       event.preventDefault();
 
-      if ((event.key === 'ArrowLeft' && !this.vertical) || (event.key === 'ArrowUp' && this.vertical)) {
+      if (
+        (event.key === 'ArrowLeft' && this.orientation === 'horizontal') ||
+        (event.key === 'ArrowUp' && this.orientation === 'vertical')
+      ) {
         newPosition -= incr;
       }
 
-      if ((event.key === 'ArrowRight' && !this.vertical) || (event.key === 'ArrowDown' && this.vertical)) {
+      if (
+        (event.key === 'ArrowRight' && this.orientation === 'horizontal') ||
+        (event.key === 'ArrowDown' && this.orientation === 'vertical')
+      ) {
         newPosition += incr;
       }
 
@@ -208,7 +214,7 @@ export default class WaSplitPanel extends WebAwesomeElement {
 
   private handleResize(entries: ResizeObserverEntry[]) {
     const { width, height } = entries[0].contentRect;
-    this.size = this.vertical ? height : width;
+    this.size = this.orientation === 'vertical' ? height : width;
 
     // There's some weird logic that gets `this.cachedPositionInPixels = NaN` or `this.position === Infinity` when
     // a split-panel goes from `display: none;` to showing.
@@ -244,8 +250,8 @@ export default class WaSplitPanel extends WebAwesomeElement {
   }
 
   render() {
-    const gridTemplate = this.vertical ? 'gridTemplateRows' : 'gridTemplateColumns';
-    const gridTemplateAlt = this.vertical ? 'gridTemplateColumns' : 'gridTemplateRows';
+    const gridTemplate = this.orientation === 'vertical' ? 'gridTemplateRows' : 'gridTemplateColumns';
+    const gridTemplateAlt = this.orientation === 'vertical' ? 'gridTemplateColumns' : 'gridTemplateRows';
     const isRtl = this.hasUpdated ? this.localize.dir() === 'rtl' : this.dir === 'rtl';
     const primary = `
       clamp(
@@ -267,13 +273,13 @@ export default class WaSplitPanel extends WebAwesomeElement {
     }
 
     if (this.primary === 'end') {
-      if (isRtl && !this.vertical) {
+      if (isRtl && this.orientation === 'horizontal') {
         this.style[gridTemplate] = `${primary} var(--divider-width) ${secondary}`;
       } else {
         this.style[gridTemplate] = `${secondary} var(--divider-width) ${primary}`;
       }
     } else {
-      if (isRtl && !this.vertical) {
+      if (isRtl && this.orientation === 'horizontal') {
         this.style[gridTemplate] = `${secondary} var(--divider-width) ${primary}`;
       } else {
         this.style[gridTemplate] = `${primary} var(--divider-width) ${secondary}`;